+++ sentinel/src/main/kotlin/com/infinum/sentinel/data/models/memory/triggers/TriggerType.kt
@@ -3,6 +3,7 @@ package com.infinum.sentinel.data.models.memory.triggers
 internal enum class TriggerType {
     MANUAL,
     SHAKE,
+    PROXIMITY,
     FOREGROUND,
     USB_CONNECTED,
     AIRPLANE_MODE_ON