+++ sentinel/src/main/kotlin/com/infinum/sentinel/data/sources/local/room/callbacks/SentinelDefaultValuesCallback.kt
@@ -9,9 +9,10 @@ internal class SentinelDefaultValuesCallback : RoomDatabase.Callback() {
         private val DEFAULT_TRIGGERS = """
         INSERT INTO "triggers" ("id","type","enabled","editable") VALUES (0,'MANUAL',1,0),
          (1,'SHAKE',1,0),
-         (2,'FOREGROUND',1,1),
-         (3,'USB_CONNECTED',1,1),
-         (4,'AIRPLANE_MODE_ON',1,1);
+         (2,'PROXIMITY',1,1),
+         (3,'FOREGROUND',1,1),
+         (4,'USB_CONNECTED',1,1),
+         (5,'AIRPLANE_MODE_ON',1,1);
         """.trimIndent()
 
         private val DEFAULT_FORMATS = """
@@ -29,6 +30,12 @@ internal class SentinelDefaultValuesCallback : RoomDatabase.Callback() {
         "fragment_arguments","fragment_saved_state") 
         VALUES (0,500,1,1,1,1,1);
         """.trimIndent()
+
+        private val DEFAULT_CRASH_MONITOR = """
+        INSERT INTO "crash_monitor" 
+        ("id","notify_exceptions","notify_anrs","include_all_data") 
+        VALUES (0,0,0,0);
+        """.trimIndent()
     }
 
     override fun onCreate(db: SupportSQLiteDatabase) {
@@ -37,6 +44,7 @@ internal class SentinelDefaultValuesCallback : RoomDatabase.Callback() {
             db.execSQL(DEFAULT_TRIGGERS)
             db.execSQL(DEFAULT_FORMATS)
             db.execSQL(DEFAULT_BUNDLE_MONITOR)
+            db.execSQL(DEFAULT_CRASH_MONITOR)
             db.setTransactionSuccessful()
             db.endTransaction()
         }